backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RAG Chatbot Backend for Physical AI and Humanoid Robotics Textbook"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "cohere>=5.5.0",
    "qdrant-client>=1.9.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.7",
]
